pyproject.toml
[project]
name = "vlad-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "cappa>=0.31.0",
    "fastapi>=0.135.3",
    "fastapi-users[oauth,sqlalchemy]>=15.0.5",
    "fastapi-users-db-sqlmodel>=0.3.0",
    "isort>=8.0.1",
    "loguru>=0.7.3",
    "msgspec>=0.20.0",
    "mypy>=1.20.0",
    "pydantic-settings>=2.13.1",
    "ruff",
    "sqlalchemy>=2.0.49",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.44.0",
    "redis[hiredis]>=7.4.0",
    "fastapi-mail>=1.6.2",
    "asgi-csrf>=0.11",
    "alembic>=1.18.4",
    "fastapi-csrf-protect>=1.0.7",
    "starlette-context>=0.5.1",
    "pyyaml>=6.0.3",
    "types-pyyaml>=6.0.12.20260408",
    "opentelemetry-api>=1.41.0",
    "opentelemetry-instrumentation-fastapi>=0.62b0",
    "opentelemetry-sdk>=1.41.0",
    "fastapi-i18n>=0.12.0",
    "fastapi-babel>=1.0.0",
    "fastapi-pagination>=0.15.12",
    "starlette-csrf>=3.0.0",
    "opentelemetry-instrumentation-sqlalchemy>=0.62b0",
    "opentelemetry-instrumentation-asyncio>=0.62b0",
    "fastcrud>=0.21.0",
]


[tool.isort]
profile = "black"
src_paths = ["backend/"]
skip = [
    "backend/alembic",
    ".venv",
]
combine_as_imports = true

[tool.ruff]
line-length = 88
exclude = ["backend/alembic"]
include = ["backend/**/*.py"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "COM",
    "C90",
    "Q",
    "W",
    "FAST001",
    "FAST002",
    "FAST003",
    "ASYNC",
    "A",
    "DTZ",
    "EM",
    "T20",
    "PT",
    "RET",
    "SIM",
    "PTH",
    "N",
]
ignore = ["COM812"]
preview = true

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
avoid-escape = true

[tool.ruff.lint.mccabe]
max-complexity = 10
